Add Base variation for ColorPickerFormCell
